dtrace: enforce inlining of dtrace_dif_variable
The dtrace_dif_variable() function is inlined during some compilations
and not during others, causing the number of frames to be skipped in
DTrace kernel stack traces to not be a constant.  That causes incorrect
values for stackdepth to be reported.

This commit requests dtrace_dif_variable() to always be inlined, and
adjusts the aframes values in function of the inlining.

Orabug: 25872472
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas jedlicka <tomas.jedlicka@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>