RDS/IB: active bonding port state fix for intfs added late
When new interfaces are added after boot or a late notifier
events cause an interface to be added late, there is need
to make sure port state moves to UP or DOWN (and does not
stay in INIT state) regardless of order of the initialization
of data structures racing with NETDEV notifier events.

Without that subsequent failover/failback processing may
not happen properly as it looks for port_state in
UP or DOWN state.

Orabug: 26139385

Signed-off-by: Mukesh Kacker <mukesh.kacker@oracle.com>
Acked-by: Wengang Wang <wen.gang.wang@oracle.com>
Signed-off-by: Brian Maly <brian.maly@oracle.com>