dtrace: touching the stack guard page is not always fatal
There is code (e.g. DTrace) that permits arbitrary memory access in
a way that provides graceful handling for failures, i.e. touching an
invalid memory location shouldn't be fatal.  By adding a notifier
call for page fault in the case of guard page access we allow handlers
to selectiely allow the access to not cause a crash.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Acked-by: Nick Alcock <nick.alcock@oracle.com>